+++ packages/jsii-pacmak/test/expected.jsii-calc/dotnet/Amazon.JSII.Tests.CalculatorPackageId/Amazon/JSII/Tests/CalculatorNamespace/MutableObjectLiteralProxy.cs
@@ -1,0 +1,19 @@
+using Amazon.JSII.Runtime.Deputy;
+
+namespace Amazon.JSII.Tests.CalculatorNamespace
+{
+    [JsiiTypeProxy(typeof(IMutableObjectLiteral), "jsii-calc.MutableObjectLiteral")]
+    internal sealed class MutableObjectLiteralProxy : DeputyBase, IMutableObjectLiteral
+    {
+        private MutableObjectLiteralProxy(ByRefValue reference): base(reference)
+        {
+        }
+
+        [JsiiProperty("value", "{\"primitive\":\"string\"}")]
+        public string Value
+        {
+            get => GetInstanceProperty<string>();
+            set => SetInstanceProperty(value);
+        }
+    }
+}