+++ packages/jsii/test/negatives/neg.property-name.1.ts
@@ -1,4 +1,4 @@
-///!MATCH_ERROR: Method and property names must use camelCase: PROP
+///!MATCH_ERROR: Method and non-static non-readonly property names must use camelCase: PROP
 
 export class MyClass {
     public PROP?: number;